flask_test1/pyproject.toml
[tool.poetry]
name = "flask-test1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"
flask-migrate = "^4.0.7"
flask-wtf = "^1.2.1"
flask-bcrypt = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
